Refactor: move bmqp::ProtocolUtil::QueueInfo to a separate source
