+++ src/groups/bmq/bmqp/bmqp_protocolutil.h
@@ -40,8 +40,6 @@
 // Thread safe.
 
 // BMQ
-
-#include <bmqc_twokeyhashmap.h>
 #include <bmqp_ctrlmsg_messages.h>
 #include <bmqp_protocol.h>
 #include <bmqt_resultcode.h>
@@ -64,10 +62,7 @@
 #include <bsl_ostream.h>
 #include <bsl_streambuf.h>
 #include <bsl_string.h>
-#include <bsl_utility.h>
 #include <bsl_vector.h>
-#include <bsla_annotations.h>
-#include <bslim_printer.h>
 #include <bslma_allocator.h>
 #include <bsls_assert.h>
 #include <bsls_performancehint.h>
@@ -82,124 +77,6 @@ namespace bmqp {
 
 /// Utilities for BlazingMQ protocol builders and iterators
 struct ProtocolUtil {
-    // PUBLIC TYPES
-
-    template <class VALUE>
-    struct QueueInfo {
-        struct StreamInfo {
-            VALUE d_value;
-
-            StreamInfo(const VALUE& value);
-        };
-
-        typedef bmqc::TwoKeyHashMap<bsl::string, unsigned int, StreamInfo>
-            StreamsMap;
-        // Map {appId, subId} to StreamInfo
-
-        /// This struct provides named access to `VALUE` - by `appId` and by
-        /// `subId`
-        struct iterator {
-            typename StreamsMap::iterator d_iterator;
-
-            iterator(const typename StreamsMap::iterator& it);
-            iterator(const iterator& other);
-
-            VALUE& value();
-
-            unsigned int subId() const;
-
-            const bsl::string& appId() const;
-
-            iterator& operator++();
-
-            iterator* operator->();
-
-            bool operator!=(const iterator& other) const;
-            bool operator==(const iterator& other) const;
-        };
-        /// This struct provides named access to `VALUE` - by `appId` and by
-        /// `subId`
-        struct const_iterator {
-            typename StreamsMap::const_iterator d_iterator;
-
-            const_iterator(const typename StreamsMap::iterator& it);
-            const_iterator(const typename StreamsMap::const_iterator& cit);
-
-            const_iterator(const iterator& other);
-
-            const VALUE& value() const;
-
-            unsigned int subId() const;
-
-            const bsl::string& appId() const;
-
-            const_iterator& operator++();
-
-            const_iterator* operator->();
-
-            bool operator!=(const const_iterator& other) const;
-            bool operator==(const const_iterator& other) const;
-        };
-
-        /// Map subscriptionId to StreamInfo
-        typedef bsl::unordered_map<unsigned int, iterator> SubscriptionsMap;
-        typedef bsl::vector<bmqp_ctrlmsg::Subscription>    Subscriptions;
-
-        StreamsMap d_streams;
-
-        SubscriptionsMap d_subscriptions;
-
-        bslma::Allocator* d_allocator_p;
-
-        QueueInfo(bslma::Allocator* allocator);
-        QueueInfo(const QueueInfo& other, bslma::Allocator* allocator = 0);
-
-        iterator insert(const bsl::string& appId,
-                        unsigned int       subId,
-                        const VALUE&       value);
-        iterator insert(const bmqp_ctrlmsg::QueueHandleParameters& parameters,
-                        const VALUE&                               value);
-        iterator insert(const bmqp_ctrlmsg::QueueStreamParameters& parameters,
-                        const VALUE&                               value);
-
-        iterator       findBySubIdSafe(unsigned int subId);
-        const_iterator findBySubIdSafe(unsigned int subId) const;
-
-        iterator       findByAppIdSafe(const bsl::string& appId);
-        const_iterator findByAppIdSafe(const bsl::string& appId) const;
-
-        iterator       findBySubId(unsigned int subId);
-        const_iterator findBySubId(unsigned int subId) const;
-
-        iterator findByHandleParameters(
-            const bmqp_ctrlmsg::QueueHandleParameters& parameters);
-
-        iterator       findByAppId(const bsl::string& appId);
-        const_iterator findByAppId(const bsl::string& appId) const;
-
-        const_iterator
-        findBySubscriptionIdSafe(unsigned int subscriptionId) const;
-
-        const_iterator findBySubscriptionId(unsigned int subscriptionId) const;
-
-        const_iterator begin() const;
-        iterator       begin();
-
-        const_iterator end() const;
-        iterator       end();
-
-        void erase(const_iterator it);
-
-        void removeSubscriptions(const bsl::string& appId);
-
-        void
-        addSubscriptions(const bmqp_ctrlmsg::StreamParameters& parameters);
-
-        void clear();
-
-        typename StreamsMap::size_type size() const;
-    };
-
     // CLASS DATA
 
     /// This public constant represents an empty appId.
@@ -857,451 +734,6 @@ ProtocolUtil::consumerInfo(const bmqp_ctrlmsg::StreamParameters& from)
 
     // Consider only the first (highest priority) consumer
     return from.subscriptions()[0].consumers()[0];
-}
-
-// ------------------------------------------
-// struct ProtocolUtil::QueueInfo::StreamInfo
-// ------------------------------------------
-
-template <class VALUE>
-inline ProtocolUtil::QueueInfo<VALUE>::StreamInfo::StreamInfo(
-    const VALUE& value)
-: d_value(value)
-{
-    // NOTHING
-}
-
-// ----------------------------------------
-// struct ProtocolUtil::QueueInfo::iterator
-// ----------------------------------------
-
-template <class VALUE>
-inline ProtocolUtil::QueueInfo<VALUE>::iterator::iterator(
-    const typename StreamsMap::iterator& it)
-: d_iterator(it)
-{
-    // NOTHING
-}
-
-template <class VALUE>
-inline ProtocolUtil::QueueInfo<VALUE>::iterator::iterator(
-    const iterator& other)
-: d_iterator(other.d_iterator)
-{
-    // NOTHING
-}
-
-template <class VALUE>
-inline VALUE& ProtocolUtil::QueueInfo<VALUE>::iterator::value()
-{
-    return d_iterator->value().d_value;
-}
-
-template <class VALUE>
-inline unsigned int ProtocolUtil::QueueInfo<VALUE>::iterator::subId() const
-{
-    return d_iterator->key2();
-}
-
-template <class VALUE>
-inline const bsl::string&
-ProtocolUtil::QueueInfo<VALUE>::iterator::appId() const
-{
-    return d_iterator->key1();
-}
-
-template <class VALUE>
-inline typename ProtocolUtil::QueueInfo<VALUE>::iterator&
-ProtocolUtil::QueueInfo<VALUE>::iterator::operator++()
-{
-    ++d_iterator;
-    return *this;
-}
-
-template <class VALUE>
-inline typename ProtocolUtil::QueueInfo<VALUE>::iterator*
-ProtocolUtil::QueueInfo<VALUE>::iterator::operator->()
-{
-    return this;
-}
-
-template <class VALUE>
-inline bool ProtocolUtil::QueueInfo<VALUE>::iterator::operator!=(
-    const iterator& other) const
-{
-    return d_iterator != other.d_iterator;
-}
-
-template <class VALUE>
-inline bool ProtocolUtil::QueueInfo<VALUE>::iterator::operator==(
-    const iterator& other) const
-{
-    return d_iterator == other.d_iterator;
-}
-
-// ----------------------------------------------
-// struct ProtocolUtil::QueueInfo::const_iterator
-// ----------------------------------------------
-
-template <class VALUE>
-inline ProtocolUtil::QueueInfo<VALUE>::const_iterator::const_iterator(
-    const typename StreamsMap::iterator& it)
-: d_iterator(it)
-{
-    // NOTHING
-}
-
-template <class VALUE>
-inline ProtocolUtil::QueueInfo<VALUE>::const_iterator::const_iterator(
-    const typename StreamsMap::const_iterator& cit)
-: d_iterator(cit)
-{
-    // NOTHING
-}
-
-template <class VALUE>
-inline ProtocolUtil::QueueInfo<VALUE>::const_iterator::const_iterator(
-    const iterator& other)
-: d_iterator(other.d_iterator)
-{
-    // NOTHING
-}
-
-template <class VALUE>
-inline const VALUE&
-ProtocolUtil::QueueInfo<VALUE>::const_iterator::value() const
-{
-    return d_iterator->value().d_value;
-}
-
-template <class VALUE>
-inline unsigned int
-ProtocolUtil::QueueInfo<VALUE>::const_iterator::subId() const
-{
-    return d_iterator->key2();
-}
-
-template <class VALUE>
-inline const bsl::string&
-ProtocolUtil::QueueInfo<VALUE>::const_iterator::appId() const
-{
-    return d_iterator->key1();
-}
-
-template <class VALUE>
-inline typename ProtocolUtil::QueueInfo<VALUE>::const_iterator&
-ProtocolUtil::QueueInfo<VALUE>::const_iterator::operator++()
-{
-    ++d_iterator;
-    return *this;
-}
-
-template <class VALUE>
-inline typename ProtocolUtil::QueueInfo<VALUE>::const_iterator*
-ProtocolUtil::QueueInfo<VALUE>::const_iterator::operator->()
-{
-    return this;
-}
-
-template <class VALUE>
-inline bool ProtocolUtil::QueueInfo<VALUE>::const_iterator::operator!=(
-    const const_iterator& other) const
-{
-    return d_iterator != other.d_iterator;
-}
-
-template <class VALUE>
-inline bool ProtocolUtil::QueueInfo<VALUE>::const_iterator::operator==(
-    const const_iterator& other) const
-{
-    return d_iterator == other.d_iterator;
-}
-
-template <class VALUE>
-inline ProtocolUtil::QueueInfo<VALUE>::QueueInfo(bslma::Allocator* allocator)
-: d_streams(allocator)
-, d_subscriptions(allocator)
-, d_allocator_p(allocator)
-{
-}
-
-template <class VALUE>
-inline ProtocolUtil::QueueInfo<VALUE>::QueueInfo(const QueueInfo&  other,
-                                                 bslma::Allocator* allocator)
-: d_streams(other.d_streams)
-, d_subscriptions(other.d_subscriptions)
-, d_allocator_p(allocator)
-{
-    // NOTHING
-}
-
-template <class VALUE>
-inline typename ProtocolUtil::QueueInfo<VALUE>::iterator
-ProtocolUtil::QueueInfo<VALUE>::insert(const bsl::string& appId,
-                                       unsigned int       subId,
-                                       const VALUE&       value)
-{
-    // Extract previous Subscriptions pointing to the 'appId'
-    bsl::vector<bsls::Types::Uint64> subscriptions(d_allocator_p);
-
-    typename SubscriptionsMap::iterator it = d_subscriptions.begin();
-    while (it != d_subscriptions.end()) {
-        iterator subQueue(it->second);
-        if (subQueue->appId() == appId) {
-            subscriptions.push_back(it->first);
-            it = d_subscriptions.erase(it);
-        }
-        else {
-            ++it;
-        }
-    }
-
-    d_streams.eraseByKey1(appId);
-
-    bsl::pair<typename StreamsMap::iterator, typename StreamsMap::InsertResult>
-        result = d_streams.insert(appId, subId, StreamInfo(value));
-
-    BSLS_ASSERT_SAFE(result.second == StreamsMap::e_INSERTED);
-
-    // Point extracted Subscriptions to the new entry
-    iterator itStream(result.first);
-
-    for (size_t i = 0; i < subscriptions.size(); ++i) {
-        BSLA_MAYBE_UNUSED bsl::pair<typename SubscriptionsMap::iterator, bool>
-                          insertRC = d_subscriptions.insert(
-                bsl::make_pair(subscriptions[i], itStream));
-        BSLS_ASSERT_SAFE(insertRC.second);
-    }
-    return itStream;
-}
-
-template <class VALUE>
-inline typename ProtocolUtil::QueueInfo<VALUE>::iterator
-ProtocolUtil::QueueInfo<VALUE>::insert(
-    const bmqp_ctrlmsg::QueueHandleParameters& parameters,
-    const VALUE&                               value)
-{
-    bmqp_ctrlmsg::SubQueueIdInfo id(d_allocator_p);
-
-    if (!parameters.subIdInfo().isNull()) {
-        id = parameters.subIdInfo().value();
-    }
-
-    return insert(id.appId(), id.subId(), value);
-}
-
-template <class VALUE>
-inline typename ProtocolUtil::QueueInfo<VALUE>::iterator
-ProtocolUtil::QueueInfo<VALUE>::insert(
-    const bmqp_ctrlmsg::QueueStreamParameters& parameters,
-    const VALUE&                               value)
-{
-    bmqp_ctrlmsg::SubQueueIdInfo id(d_allocator_p);
-
-    if (!parameters.subIdInfo().isNull()) {
-        id = parameters.subIdInfo().value();
-    }
-
-    return insert(id.appId(), id.subId(), value);
-}
-
-template <class VALUE>
-inline typename ProtocolUtil::QueueInfo<VALUE>::iterator
-ProtocolUtil::QueueInfo<VALUE>::findBySubIdSafe(unsigned int subId)
-{
-    return iterator(d_streams.findByKey2(subId));
-}
-
-template <class VALUE>
-inline typename ProtocolUtil::QueueInfo<VALUE>::const_iterator
-ProtocolUtil::QueueInfo<VALUE>::findBySubIdSafe(unsigned int subId) const
-{
-    return const_iterator(d_streams.findByKey2(subId));
-}
-
-template <class VALUE>
-inline typename ProtocolUtil::QueueInfo<VALUE>::iterator
-ProtocolUtil::QueueInfo<VALUE>::findByAppIdSafe(const bsl::string& appId)
-{
-    return iterator(d_streams.findByKey1(appId));
-}
-
-template <class VALUE>
-inline typename ProtocolUtil::QueueInfo<VALUE>::const_iterator
-ProtocolUtil::QueueInfo<VALUE>::findByAppIdSafe(const bsl::string& appId) const
-{
-    return const_iterator(d_streams.findByKey1(appId));
-}
-
-template <class VALUE>
-inline typename ProtocolUtil::QueueInfo<VALUE>::iterator
-ProtocolUtil::QueueInfo<VALUE>::findBySubId(unsigned int subId)
-{
-    iterator result = findBySubIdSafe(subId);
-    BSLS_ASSERT_SAFE(result != d_streams.end());
-
-    return result;
-}
-
-template <class VALUE>
-inline typename ProtocolUtil::QueueInfo<VALUE>::const_iterator
-ProtocolUtil::QueueInfo<VALUE>::findBySubId(unsigned int subId) const
-{
-    const_iterator result = findBySubIdSafe(subId);
-    BSLS_ASSERT_SAFE(result != d_streams.end());
-
-    return result;
-}
-
-template <class VALUE>
-inline typename ProtocolUtil::QueueInfo<VALUE>::iterator
-ProtocolUtil::QueueInfo<VALUE>::findByHandleParameters(
-    const bmqp_ctrlmsg::QueueHandleParameters& parameters)
-{
-    bmqp_ctrlmsg::SubQueueIdInfo id(d_allocator_p);
-
-    if (!parameters.subIdInfo().isNull()) {
-        id = parameters.subIdInfo().value();
-    }
-    return iterator(d_streams.findByKey2(id.subId()));
-}
-
-template <class VALUE>
-inline typename ProtocolUtil::QueueInfo<VALUE>::iterator
-ProtocolUtil::QueueInfo<VALUE>::findByAppId(const bsl::string& appId)
-{
-    iterator result = findByAppIdSafe(appId);
-    BSLS_ASSERT_SAFE(result != d_streams.end());
-
-    return result;
-}
-
-template <class VALUE>
-inline typename ProtocolUtil::QueueInfo<VALUE>::const_iterator
-ProtocolUtil::QueueInfo<VALUE>::findByAppId(const bsl::string& appId) const
-{
-    const_iterator result = findByAppIdSafe(appId);
-    BSLS_ASSERT_SAFE(result != d_streams.end());
-
-    return result;
-}
-
-template <class VALUE>
-inline typename ProtocolUtil::QueueInfo<VALUE>::const_iterator
-ProtocolUtil::QueueInfo<VALUE>::findBySubscriptionIdSafe(
-    unsigned int subscriptionId) const
-{
-    typename SubscriptionsMap::const_iterator cit = d_subscriptions.find(
-        subscriptionId);
-    return cit == d_subscriptions.end() ? end() : const_iterator(cit->second);
-}
-
-template <class VALUE>
-inline typename ProtocolUtil::QueueInfo<VALUE>::const_iterator
-ProtocolUtil::QueueInfo<VALUE>::findBySubscriptionId(
-    unsigned int subscriptionId) const
-{
-    typename SubscriptionsMap::const_iterator cit = d_subscriptions.find(
-        subscriptionId);
-
-    BSLS_ASSERT_SAFE(cit != d_subscriptions.end());
-
-    return const_iterator(cit->second);
-}
-template <class VALUE>
-inline typename ProtocolUtil::QueueInfo<VALUE>::const_iterator
-ProtocolUtil::QueueInfo<VALUE>::begin() const
-{
-    return const_iterator(d_streams.begin());
-}
-
-template <class VALUE>
-inline typename ProtocolUtil::QueueInfo<VALUE>::iterator
-ProtocolUtil::QueueInfo<VALUE>::begin()
-{
-    return iterator(d_streams.begin());
-}
-
-template <class VALUE>
-inline typename ProtocolUtil::QueueInfo<VALUE>::const_iterator
-ProtocolUtil::QueueInfo<VALUE>::end() const
-{
-    return const_iterator(d_streams.end());
-}
-
-template <class VALUE>
-inline typename ProtocolUtil::QueueInfo<VALUE>::iterator
-ProtocolUtil::QueueInfo<VALUE>::end()
-{
-    return iterator(d_streams.end());
-}
-
-template <class VALUE>
-inline void ProtocolUtil::QueueInfo<VALUE>::erase(const_iterator it)
-{
-    removeSubscriptions(it->appId());
-
-    d_streams.erase(it.d_iterator);
-}
-
-template <class VALUE>
-inline void
-ProtocolUtil::QueueInfo<VALUE>::removeSubscriptions(const bsl::string& appId)
-{
-    typename SubscriptionsMap::iterator it = d_subscriptions.begin();
-    while (it != d_subscriptions.end()) {
-        iterator subQueue(it->second);
-        if (subQueue->appId() == appId) {
-            it = d_subscriptions.erase(it);
-        }
-        else {
-            ++it;
-        }
-    }
-}
-
-template <class VALUE>
-inline void ProtocolUtil::QueueInfo<VALUE>::addSubscriptions(
-    const bmqp_ctrlmsg::StreamParameters& parameters)
-{
-    removeSubscriptions(parameters.appId());
-
-    iterator itStream = findByAppId(parameters.appId());
-
-    const Subscriptions& subscriptions = parameters.subscriptions();
-    for (Subscriptions::const_iterator it = subscriptions.begin();
-         it != subscriptions.end();
-         ++it) {
-        BSLA_MAYBE_UNUSED bsl::pair<typename SubscriptionsMap::iterator, bool>
-                          insertRC = d_subscriptions.insert(
-                bsl::make_pair(it->sId(), itStream));
-        BSLS_ASSERT_SAFE(insertRC.second);
-    }
-}
-
-template <class VALUE>
-inline void ProtocolUtil::QueueInfo<VALUE>::clear()
-{
-    d_streams.clear();
-    d_subscriptions.clear();
-}
-
-template <class VALUE>
-inline typename ProtocolUtil::QueueInfo<VALUE>::StreamsMap::size_type
-ProtocolUtil::QueueInfo<VALUE>::size() const
-{
-    return d_streams.size();
-}
-
-template <class T>
-inline bsl::ostream& operator<<(bsl::ostream&                     os,
-                                const ProtocolUtil::QueueInfo<T>& rhs)
-{
-    bslim::Printer printer(&os, 0, -1);  // one line
-
-    printer.printValue(rhs.begin(), rhs.end());
-
-    return os;
 }
 
 }  // close package namespace
